requirements.txt
streamlit
streamlit_option_menu
gspread
smtplib
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
